+++ src/main/kotlin/cloud/glitchdev/rfu/constants/fishing/TrophyFrog.kt
@@ -1,18 +1,25 @@
 package cloud.glitchdev.rfu.constants.fishing
 
-enum class TrophyFrog(override val displayName: String) : Trophy {
-    COMMON_FROG("Common Frog"),
-    LEAP_FROG("Leap Frog"),
-    WETLANDS_FROG("Wetlands Frog"),
-    REALITY_HOPPER("Reality Hopper"),
-    EXPLODING_FROG("Exploding Frog"),
-    SEA_FROG("Sea Frog"),
-    BLESSED_FROG("Blessed Frog"),
-    TREE_FROG("Tree Frog"),
-    CAVE_FROG("Cave Frog"),
-    HIGHLANDS_FROG("Highlands Frog"),
-    BULLFROG("Bullfrog"),
-    PUDDLE_JUMPER("Puddle Jumper");
+import cloud.glitchdev.rfu.constants.skyblock.Rarity
+
+enum class TrophyFrog(
+    override val displayName: String,
+    override val rarity: Rarity = Rarity.COMMON,
+    override val goldPity: Int = 100,
+    override val diamondPity: Int = 600
+) : Trophy {
+    COMMON_FROG("Common Frog", Rarity.COMMON),
+    LEAP_FROG("Leap Frog", Rarity.UNCOMMON),
+    WETLANDS_FROG("Wetlands Frog", Rarity.UNCOMMON),
+    REALITY_HOPPER("Reality Hopper", Rarity.UNCOMMON),
+    EXPLODING_FROG("Exploding Frog", Rarity.UNCOMMON, 50, 300),
+    SEA_FROG("Sea Frog", Rarity.RARE),
+    BLESSED_FROG("Blessed Frog", Rarity.RARE),
+    BULLFROG("Bullfrog", Rarity.RARE),
+    TREE_FROG("Tree Frog", Rarity.EPIC),
+    CAVE_FROG("Cave Frog", Rarity.EPIC),
+    HIGHLANDS_FROG("Highlands Frog", Rarity.EPIC),
+    PUDDLE_JUMPER("Puddle Jumper", Rarity.LEGENDARY, 50, 300);
 
     companion object {
         fun fromName(name: String): TrophyFrog? {