+++ src/main/kotlin/cloud/glitchdev/rfu/data/fishing/TrophyPityEntry.kt
@@ -1,0 +1,7 @@
+package cloud.glitchdev.rfu.data.fishing
+
+data class TrophyPityEntry(
+    val name: String,
+    var goldProgress: Int = 0,
+    var diamondProgress: Int = 0
+)